# docs/openapi/trivida-openapi.yaml
openapi: 3.0.3
info:
  title: Trivida API
  version: 1.0.0
  description: >-
    Backend REST API for Trivida MVP (Node/Express + MongoDB). JWT access 15 min, refresh rotation 7 days.
servers:
  - url: https://api.trivida.example.com
    description: Production
  - url: http://localhost:4000
    description: Local Dev
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CronKey:
      type: apiKey
      in: header
      name: X-Cron-Key
...
  schemas:
...
    FXRates:
      type: object
      properties:
        base: { type: string }
        rates:
          type: object
          additionalProperties: { type: number }
        fetchedAt: { type: string, format: date-time }
        ttlSeconds: { type: integer }
    Weather:
      type: object
      properties:
        condition: { type: string }
        temperature: { type: number }
        humidity: { type: number }
        city: { type: string }
        country: { type: string }
        fetchedAt: { type: string, format: date-time }
        ttlSeconds: { type: integer }
    Timezone:
      type: object
      properties:
        zoneName: { type: string }
        gmtOffset: { type: integer }
        countryName: { type: string }
        fetchedAt: { type: string, format: date-time }
        ttlSeconds: { type: integer }
...
paths:
...
  /fx/rates:
    get:
      tags: [External]
      parameters:
        - name: base
          in: query
          schema: { type: string, default: XAF }
        - name: symbols
          in: query
          schema: { type: string, example: USD,EUR }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FXRates' } } } }
  /weather:
    get:
      tags: [External]
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
        - name: units
          in: query
          schema: { type: string, enum: [metric, imperial], default: metric }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Weather' } } } }
  /timezone:
    get:
      tags: [External]
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Timezone' } } } }
...